requirements.txt
python-dotenv
requests
bs4
mysql-connector-python